requirements.txt
torch==2.5.0
torchvision==0.20.0
torchaudio==2.5.0
PyYAML==6.0.3
scipy==1.15.3
tqdm==4.67.1
mlconfig==0.3.2
librosa==0.11.0
Pillow==12.0.0
numpy==2.2.6
matplotlib==3.10.7
einops==0.8.1
toml==0.10.2
pyloudnorm==0.1.1